requirements.txt
langchain
langchain-openai
flask
flask_cors
faiss-cpu
openai
PyPDF2
python-dotenv
pandas
werkzeug==0.10.4